requirements_bot.txt
# Trading Bot Requirements
# ========================
# Minimal dependencies for the trading bot
# Use this for Docker deployments

# Core
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Alpaca Trading API
alpaca-py>=0.20.0

# Optional: If using requests-based API calls
requests>=2.28.0

# Logging
# (stdlib logging is sufficient, no extra deps needed)
